+++ objectified-ui/public/WHATS_NEW.md
@@ -16,6 +16,7 @@ We continue to improve the platform based on your feedback with improvements and
 - Import: Publish uses short note via description, title, or slug and version if not provided
 - UI: Fixes Published Versions viewing of OpenAPI and Arazzo URLs
 - UI: Fixes Published Versions list so versions starting with a "v" don't duplicate and show "vv"
+- DB: Optimizations made to increase speed for import processing
 
 ---